findmyfuel/app/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "findmyfuel"
version = "0.1.0"
description = "Laptop prototype API for querying cheap nearby fuel from the UK Fuel Finder service."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115,<1.0",
  "httpx>=0.28,<1.0",
  "python-dotenv>=1.1,<2.0",
  "uvicorn>=0.34,<1.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
